Basic organization, simple configuration.
This simple configuration has the smallest number of changes that I
can use for a working emacs system, assuming emacs 24.

I'm going to use this as a base to work off of, and slowly pull in
configurations that I've got spread out over many different installs
of emacs on many different computers.