Re-arranged init.el, added some documentation.
It makes sense to have the smaller customizations at the top of the
file. This way, if something breaks I still have ido mode and backups
are still saved outside of the current directory.

Also, having them at the top makes them more visible, and more likely
to be re-factored if anything gets out of control.