Added package control code, don't track packages.
Emacs 24 comes with a native package manager. In this commit I'm adding
the Marmalade repository to the default repository, and setting it up
to autoload all of the packages I've installed.

Packages installed from repositories will not be tracked here, however.
They will just be assumed to be in working order.