Adds function to install uninstalled packages.
In the init.el file there is now a variable that has a list called
`keep-installed`. At startup, emacs will check if all of these
packages are installed. If any of them are missing, the user will be
prompted to install them.

This code may not be sensitive to versions of these package (I'm not
sure exactly how emacs's package managment controls versions) so if
future research shows this not to be safe, this code could use some
refactoring.