Added defuns.el, dollar pair binding.
The defuns.el file is created to hold some functions that I define. It
is set up to be the first of my files to load in my `init.el` file, so
it should be available when the rest of the files load.

I also define a funciton `insert-dollar-pair` which inserts two dollar
signs and moves the point between them. This is bound to "C-c 4" and
is usefull to enter inline-math mode in LaTeX.